+++ pyron-engine/src/main/scala/com/cloudentity/pyron/api/ApiHandler.scala
@@ -5,6 +5,7 @@ import com.cloudentity.pyron.api.body.{BodyBuffer, BodyLimit, RequestBodyTooLarg
 import com.cloudentity.pyron.apigroup.{ApiGroup, ApiGroupConf, ApiGroupsChangeListener, ApiGroupsStore}
 import com.cloudentity.pyron.client.{TargetClient, TargetResponse}
 import com.cloudentity.pyron.config.Conf
+import com.cloudentity.pyron.config.Conf.AppConf
 import com.cloudentity.pyron.domain.flow.{FlowFailure, _}
 import com.cloudentity.pyron.domain.http._
 import com.cloudentity.pyron.domain.rule._
@@ -34,7 +35,7 @@ import scala.util.{Failure, Success, Try}
 
 trait ApiHandler {
   @VertxEndpoint
-  def handle(defaultRequestBodyMaxSize: Option[Kilobytes], ctx: RoutingContext): VxFuture[Unit]
+  def handle(conf: AppConf, ctx: RoutingContext): VxFuture[Unit]
 }
 
 class ApiHandlerVerticle extends ScalaServiceVerticle with ApiHandler with ApiGroupsChangeListener {
@@ -77,7 +78,7 @@ class ApiHandlerVerticle extends ScalaServiceVerticle with ApiHandler with ApiGr
         targetClient = newTargetClient
       }
 
-  def handle(defaultRequestBodyMaxSize: Option[Kilobytes], ctx: RoutingContext): VxFuture[Unit] = {
+  def handle(conf: AppConf, ctx: RoutingContext): VxFuture[Unit] = {
     val vertxRequest   = ctx.request()
     val vertxResponse  = ctx.response()
     val tracingContext = RoutingWithTracingS.getOrCreate(ctx, getTracing)
@@ -107,7 +108,9 @@ class ApiHandlerVerticle extends ScalaServiceVerticle with ApiHandler with ApiGr
 
           ApiRequestHandler.setRule(ctx, rule)
           setTracingOperationName(ctx, rule)
-          log.debug(tracingContext, s"Found $rule for, request='$requestSignature'")
+
+          log.debug(tracingContext, s"Found ${ruleWithPathParams.rule} for, request='$requestSignature'")
+          val defaultRequestBodyMaxSize = conf.defaultRequestBodyMaxSize
 
           for {
 
@@ -150,6 +153,9 @@ class ApiHandlerVerticle extends ScalaServiceVerticle with ApiHandler with ApiGr
       }
 
     program.onComplete { result =>
+      if (conf.traceIdHeaderEnabled.contains(true)) {
+        vertxResponse.putHeader("Trace-Id", tracingContext.getTraceId)
+      }
       try {
         result match {
           case Success(apiResponse) =>